chore: Automate adding clog to manifest (#208)
Context:

In c886756b2783d1289eaaf96b950c05d02585e26b a clog generation script was
added in order to help automate changelog generation. However, generated
clogs still needed to be manually added to manifest.json. At the time,
use of jq was experimented with in the create-clog.sh script, but was
found to be not worth the effort.

Summary:

Adds a NodeJS script written in TS for adding changelog entries to
manifest.json.

Details:

The manifest.json file was formatted as part of this diff because the
add_clog_to_manifest.ts script uses JSON.stringify to serialize changes
to JSON. The resulting format differed from our prior format, but still
complies with prettier.