feat(cdr): document disabling built-in auth
Summary:

As of release 1.27, built-in authentication can be disabled if and only
if OIDC is configured. The UI will not display the option for using
built-in authentication and the API will return with an error response.
Site managers may still view the option using a query parameter and will
be allowed to authenticate this way.